requirements.txt
python-telegram-bot==20.0
python-dotenv==1.0.0
requests==2.31.0
pillow>=10.0.0
schedule==1.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
